% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_xdf.R
\name{arrange_.RxFileData}
\alias{arrange}
\alias{arrange_}
\alias{arrange_.RxFileData}
\title{Arrange the rows in an Xdf file}
\usage{
\method{arrange_}{RxFileData}(.data, ..., .output, .rxArgs, .dots)
}
\arguments{
\item{...}{List of unquoted variable names. Use \code{desc} to sort in descending order.}

\item{.output}{Output format for the returned data. If not supplied, create an xdf tbl; if \code{NULL}, return a data frame; if a character string naming a file, save an Xdf file at that location.}

\item{.rxArgs}{A list of RevoScaleR arguments. See \code{\link{rxArgs}} for details.}

\item{.dots}{Used to work around non-standard evaluation. See the dplyr documentation for details.}

\item{data}{An Xdf data source, tbl, or other RevoScaleR file data source.}
}
\value{
An object representing the sorted data. This depends on the \code{.output} argument: if missing, it will be an xdf tbl object; if \code{NULL}, a data frame; and if a filename, an Xdf data source referencing a file saved to that location.

\code{\link{rxSort}}, \code{\link[dplyr]{arrange}} in package dplyr
}
\description{
Arrange the rows in an Xdf file
}
\details{
The underlying RevoScaleR function is \code{rxSort}. This has many sorting options, including removing duplicated keys, adding a column of frequency counts, and so on.
}

