% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tbl_xdf.R
\name{as.data.frame.RxXdfData}
\alias{as.data.frame.RxFileData}
\alias{as.data.frame.RxXdfData}
\title{Convert a data source or tbl to a data frame}
\usage{
\method{as.data.frame}{RxXdfData}(x, ...)

\method{as.data.frame}{RxFileData}(x, ...)
}
\arguments{
\item{x}{A data source object, or tbl wrapping the same.}

\item{...}{Other arguments to \code{rxReadXdf} or \code{rxImport}, as appropriate.}
}
\description{
Convert a data source or tbl to a data frame
}
\details{
This is a simple wrapper around \code{\link[RevoScaleR]{rxReadXdf}} for the \code{RxXdfData} method, and \code{\link[RevoScaleR]{rxImport}} for the \code{RxFileData} method.
}

