% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_xdf.R
\name{collect.tbl_xdf}
\alias{collect.tbl_xdf}
\title{Select columns in a data source}
\usage{
\method{collect}{tbl_xdf}(.data)
}
\arguments{
\item{.data}{A data source.}
}
\value{
An object of class "tbl_df", "tbl", "data.frame"
}
\description{
Select columns in a data source
}
\details{
All the special functions mentioned in the documentation for \code{\link[dplyr]{select}} will also work with dplyrXdf. Note that renaming a variable is \emph{not} supported in dplyrXdf \code{select}. If you want to do this, follow the \code{select} with a \code{rename}; the latter is very fast, as it only modifies the metadata portion of an xdf file, not the data itself.
Collect brings data back into an R data.frame (stored in a tbl_df).
}
\examples{
mtcars <- rxDataStep(mtcars, "mtcars.xdf", overwrite=TRUE)
class(mtcars)
out <- mtcars \%>\% 
  group_by(cyl) \%>\%
  summarise(avg_mpg = mean(mpg)) \%>\%
  collect()
out
class(out)
}
\seealso{
\code{\link[dplyr]{collect}} in package dplyr
}

