% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_xdf.R
\name{delete_xdf_tbls}
\alias{delete_xdf_tbls}
\title{Delete data files for xdf tbls}
\usage{
delete_xdf_tbls(path, fileSystem = rxGetFileSystem())
}
\arguments{
\item{path}{Directory where the files are located. Defaults to the temporary directory where dplyrXdf creates tbl files (the R temporary directory for the native filesystem, or the dplyrXdf-created directory on HDFS).}
}
\description{
Delete data files for xdf tbls
}
\details{
This is a utility function to delete the files generated by dplyrXdf. Note that all xdf files in the specified location will be removed!
}
