% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_utils.R
\name{deleteTbl}
\alias{deleteTbl}
\title{deleteTbl}
\usage{
deleteTbl(xdf)
}
\arguments{
\item{xdf}{A character string specifying the table to drop}
}
\value{
Not sure
}
\description{
Drops XDF tables
}
\details{
This was nto exported so it could not be seen
}
\author{
Jared P. Lander
}

