% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/do_xdf.R
\name{do_.RxFileData}
\alias{do}
\alias{doXdf}
\alias{doXdf_}
\alias{doXdf_.RxFileData}
\alias{doXdf_.grouped_tbl_xdf}
\alias{do_.RxFileData}
\alias{do_.grouped_tbl_xdf}
\title{Do arbitrary operations on a tbl}
\usage{
\method{do_}{RxFileData}(.data, ..., .dots)

doXdf(.data, ...)

doXdf_(.data, ..., .dots)

\method{doXdf_}{RxFileData}(.data, ..., .dots)

\method{do_}{grouped_tbl_xdf}(.data, ..., .dots)

\method{doXdf_}{grouped_tbl_xdf}(.data, ..., .dots)
}
\arguments{
\item{.data}{A tbl for an Xdf data source; or a raw Xdf data source.}

\item{...}{Expressions to apply.}

\item{.dots}{Used to work around non-standard evaluation. See the dplyr vignettes for details.}
}
\value{
The \code{do} and \code{doXdf} verbs always return a data frame, unlike the other verbs for Xdf objects.
}
\description{
The \code{do} verb converts the data to a data frame before running the operations. The \code{doXdf} verb keeps the data in Xdf format, so is not limited by memory.
}
\details{
The \code{do} verb supports named and unnamed arguments, as per \code{dplyr::do}. The \code{doXdf} verb only supports named arguments.

To run expressions on a grouped Xdf tbl, \code{do} and \code{doXdf} create a factor variable \code{.group.} containing all observed combinations of the grouping variables. The data is then split into multiple files by this factor, and the arguments are called on each file. This ensures that the groups will be appropriately generated regardless of the types of the grouping variables. Note however that this may be slow if you have a large number of groups; and, for \code{do}, the operation will be limited by memory if the number of rows per group is large.
}
\seealso{
\code{\link[dplyr]{do}} in package dplyr
}

