% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/factorise_df.R, R/factorise_xdf.R
\name{factorise_.data.frame}
\alias{factorise}
\alias{factorise_}
\alias{factorise_.RxFileData}
\alias{factorise_.RxXdfData}
\alias{factorise_.data.frame}
\alias{factorize}
\alias{factorize_}
\title{Convert columns in an Xdf file to factor}
\usage{
\method{factorise_}{data.frame}(.data, ..., .dots)

factorise(.data, ...)

factorize(.data, ...)

factorise_(.data, ..., .dots)

factorize_(.data, ..., .dots)

\method{factorise_}{RxXdfData}(.data, ..., .dots)

\method{factorise_}{RxFileData}(.data, ..., .dots)
}
\arguments{
\item{.data}{A data source.}

\item{...}{Variables to convert to factors.}

\item{.dots}{Used to work around non-standard evaluation. See the dplyr vignettes for details.}
}
\description{
Convert columns in an Xdf file to factor
}
\details{
The selector functions listed in \code{\link[dplyr]{select}} also work with \code{factorise}. In addition, you can use the following:
\itemize{
 \item \code{all_character()}: selects all character variables
 \item \code{all_integer()}: selects all integer variables, ie those of type \code{"logical"} and \code{"integer"}
 \item \code{all_numeric()}: selects all numeric variables, ie those of type \code{"numeric"}, \code{"Date"}, \code{"POSIXct"}, \code{"logical"} and \code{"integer"}
}
If no variables are specified, all character variables will be converted to factors.

You can specify the levels for a variable by specifying them as the value of the argument. For example, \code{factorise(*, x = c("a","b","c"))} will turn the variable \code{x} into a factor with three levels \code{a}, \code{b} and \code{c}. Any values that don't match the set of levels will be turned into NAs. In particular, this means you should include the existing levels for variables that are already factors.

For performance reasons, factors created by \code{factorise} are not sorted; instead, the ordering of their levels will be determined by the order in which they are encountered in the data.

The method for \code{RxXdfData} objects is a shell around \code{\link[RevoScaleR]{rxFactors}}, which is the standard RevoScaleR function for factor manipulation. For \code{RxFileData} objects, the method calls \code{\link[RevoScaleR]{rxImport}} with an appropriately constructed \code{colInfo} argument.
}
\seealso{
\code{\link[RevoScaleR]{rxFactors}}, \code{\link[RevoScaleR]{rxImport}}, \code{\link{factor}}

\code{\link{chol}}, \code{\link{qr}}, \code{\link{svd}} for the other meaning of factorise
}

