% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R, R/zzz_non_xdf_srcs.R
\name{filter.RxFileData}
\alias{filter.RxFileData}
\alias{filter}
\alias{filter.RxDataSource}
\title{Subset a data source by rows}
\usage{
\method{filter}{RxFileData}(.data, ..., .outFile = tbl_xdf(.data), .rxArgs)

\method{filter}{RxDataSource}(.data, ...)
}
\arguments{
\item{.data}{A data source.}

\item{...}{Expressions to filter by.}

\item{.outFile}{Output format for the returned data. If not supplied, create an xdf tbl; if \code{NULL}, return a data frame; if a character string naming a file, save an Xdf file at that location.}

\item{.rxArgs}{A list of RevoScaleR arguments. See \code{\link{rxArgs}} for details.}
}
\value{
An object representing the filtered data. This depends on the \code{.outFile} argument: if missing, it will be an xdf tbl object; if \code{NULL}, a data frame; and if a filename, an Xdf data source referencing a file saved to that location.
}
\description{
Subset a data source by rows
}
\examples{
mtx <- as_xdf(mtcars, overwrite=TRUE)
tbl <- filter(mtx, mpg > 20)
nrow(tbl)
tbl2 <- filter(mtx, am == 1, vs == 1)
nrow(tbl2)

# filter on rows and columns simultaneously with .rxArgs
tbl3 <- filter(mtx, am == 1, vs == 1, .rxArgs=list(varsToKeep=c("mpg", "cyl")))
dim(tbl3)

# save to a persistent Xdf file
filter(mtx, am == 1, vs == 1, .outFile="mtcars_filter.xdf")
}
\seealso{
\code{\link[dplyr]{filter}} in package dplyr, \code{\link{subset}}
}
