% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfs_utils.R
\name{local_exec}
\alias{local_exec}
\title{Runs an expression in the local compute context}
\usage{
local_exec(expr, context = "local")
}
\arguments{
\item{expr}{An expression to execute. Normally something that depends on the compute context, such as \code{rxDataStep}.}

\item{context}{The compute context in which to execute \code{expr}. Defaults to local.}
}
\value{
The value of \code{expr}.
}
\description{
Runs an expression in the local compute context
}
\details{
This function is useful when you are working with datasets in both the native filesystem and HDFS. The workhorse RevoScaleR function for data transformation, \code{rxDataStep}, will complain if you are in a distributed compute context such as \code{\link{RxHadoopMR}} or \code{\link{RxSpark}}, and you want to process a dataset in the native filesystem. You can wrap your code inside a \code{local_exec} call to switch to the local compute context temporarily, and then switch back when it has finished running.
}
\seealso{
\code{\link{eval}}
}
