% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutate_xdf.R, R/transmute_xdf.R
\name{mutate_.RxFileData}
\alias{mutate}
\alias{mutate_.RxFileData}
\alias{transmute}
\alias{transmute_.RxFileData}
\alias{transmute_.grouped_tbl_xdf}
\title{Add or modify variables}
\usage{
\method{mutate_}{RxFileData}(.data, ..., .dots)

\method{transmute_}{RxFileData}(.data, ..., .dots)

\method{transmute_}{grouped_tbl_xdf}(.data, ..., .dots)
}
\arguments{
\item{.data}{A tbl for an Xdf data source; or a raw Xdf data source.}

\item{...}{Variables to add or modify.}

\item{.dots}{Used to work around non-standard evaluation. See the dplyr vignettes for details. See "Details" below for information on passing \code{\link[RevoScaleR]{rxTransform}} arguments.}
}
\description{
Use \code{mutate} to add new variables and preserve existing ones; use \code{transmute} to keep only new and modified variables.
}
\details{
These functions call \code{\link[RevoScaleR]{rxDataStep}} to do the variable transformations. For simple transformations, namely those that might be done using \code{rxDataStep}'s \code{transforms} argument, you can simply pass these as named arguments in the main \code{mutate} or \code{transmute} call. More complex transformations can be passed in a \code{.rxArgs} argument, which should be a named list containing one or more of the \code{transformFunc}, \code{transformVars}, \code{transformObjects}, \code{transformPackages} and \code{transformEnvir} parameters.

Note that if you supply a \code{transformFunc}, its returned variables will override any transformations in the main call to \code{mutate} and \code{transmute}). In particular, the results of any such inline transformations will be lost unless you also include them in the output of the \code{transformFunc}. This mirrors the existing behaviour of the variable transformation functionality in RevoScaleR. It's not recommended to use both inline transformations and a \code{transformFunc} at the same time, as the results may be confusing.

To modify a grouped Xdf tbl, these functions create a factor variable \code{.group.} containing all observed combinations of the grouping variables. They then use \code{\link{rxSplit}} to split the data into multiple files by this factor, and call \code{rxDataStep} on each. This ensures two things: first, that the groups will be appropriately generated regardless of the types of the grouping variables; and second, the code remains scalable to large dataset sizes. Note however that this may be slow if you have a large number of groups.
}
\seealso{
\code{\link[dplyr]{mutate}} and \code{\link{transmute}} in package dplyr, \code{\link[RevoScaleR]{rxDataStep}}, \code{\link[RevoScaleR]{rxTransform}} for variable transformations in RevoScaleR
}

