% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_utils.R
\name{newTbl}
\alias{newTbl}
\title{newTbl}
\usage{
newTbl(xdf = NULL, fileSystem = rxGetFileSystem(xdf))
}
\arguments{
\item{xdf}{Name of xdf}

\item{fileSystem}{The file system we are on}
}
\value{
Not sure
}
\description{
Makes a new tbl
}
\details{
Wasn't exported and that was causing issues
}
\author{
Jared P. Lander
}

