% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R, R/zzz_non_xdf_srcs.R
\name{rename.RxFileData}
\alias{rename.RxFileData}
\alias{rename}
\alias{rename.RxDataSource}
\title{Rename columns in an Xdf file}
\usage{
\method{rename}{RxFileData}(.data, ..., .outFile = tbl_xdf(.data), .rxArgs)

\method{rename}{RxDataSource}(.data, ...)
}
\arguments{
\item{.data}{A data source.}

\item{...}{Key-value pairs of unquoted variables to rename, in new = old format.}

\item{.outFile}{Output format for the returned data. If not supplied, create an xdf tbl; if \code{NULL}, return a data frame; if a character string naming a file, save an Xdf file at that location.}

\item{.rxArgs}{A list of RevoScaleR arguments. See \code{\link{rxArgs}} for details.}
}
\value{
An object representing the transformed data. This depends on the \code{.outFile} argument: if missing, it will be an xdf tbl object; if \code{NULL}, a data frame; and if a filename, an Xdf data source referencing a file saved to that location.
}
\description{
Rename columns in an Xdf file
}
\details{
Renaming is generally a very fast operation, as only the metadata in the Xdf file header is changed. The exception is when the dataset needs to be copied, either to preserve dplyrXdf's semantics of never modifying the source data for a pipeline, or if the the desired output format is a data frame or persistent Xdf file. If a copy is made, \code{rxDataStep} is used and will accept arguments supplied in \code{.rxArgs}.
}
\examples{
mtx <- as_xdf(mtcars, overwrite=TRUE)
tbl <- rename(mtx, mpg2=mpg, cyl2=cyl)
names(tbl)

# save to a persistent Xdf file
rename(mtx, mpg2=mpg, .outFile="mtcars_renamed.xdf")
}
\seealso{
\code{\link[dplyr]{rename}} in package dplyr
}
