% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_n.RxXdfData}
\alias{sample_n.RxXdfData}
\alias{sample_frac.RxXdfData}
\alias{sample_n.grouped_tbl_xdf}
\alias{sample_frac.grouped_tbl_xdf}
\title{Do random sampling from an Xdf file}
\usage{
\method{sample_n}{RxXdfData}(tbl, size = 1, replace = FALSE,
  weight = NULL, .env = NULL)

\method{sample_frac}{RxXdfData}(tbl, size = 1, replace = FALSE,
  weight = NULL, .env = NULL)

\method{sample_n}{grouped_tbl_xdf}(tbl, size = 1, replace = FALSE,
  weight = NULL, .env = NULL)

\method{sample_frac}{grouped_tbl_xdf}(tbl, size = 1, replace = FALSE,
  weight = NULL, .env = NULL)
}
\arguments{
\item{tbl}{An Xdf file or a tbl wrapping the same.}

\item{size}{For \code{sample_n}, the number of rows to select. For \code{sample_frac}, the fraction of rows to select. For a grouped dataset, \code{size} applies to each group.}

\item{replace, weight, .env}{Not used.}
}
\value{
An Xdf tbl.
}
\description{
Do random sampling from an Xdf file
}
\details{
Sampling from Xdf files is slightly more limited than the data frame case. Only unweighted sampling without replacement is supported, and attempts to specify otherwise will result in a warning. Unlike the other single-table dplyr verbs, \code{sample_n} and \code{sample_frac} do not delete tbl inputs; this is because it's unlikely that a sample is intended to replace the input data entirely.

Currently sampling on HDFS data works in the local compute context (on the edge node) but not in the Hadoop or Spark compute contexts.
}
\seealso{
\code{\link[dplyr]{sample_frac}}, \code{\link[dplyr]{sample_n}}, \code{\link{sample}}
}
