% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_frac.R
\name{sample_frac.tbl_xdf}
\alias{sample_frac.tbl_xdf}
\title{Random sample from a xdf object}
\usage{
\method{sample_frac}{tbl_xdf}(.data, size = 1, replace = FALSE,
  weight = NULL, ...)
}
\arguments{
\item{.data}{A tbl for an Xdf data source; or a raw Xdf data source.}

\item{size}{the fraction of rows to sample}

\item{replace}{default to FALSE. \code{replace=TRUE} now allowed when working on chunks of data.}

\item{weight}{default to NULL. weighted sampling now allowed when working on chunks of data.}

\item{...}{other parameters to be passed to \code{RevoScaleR::rxDataStep}}
}
\value{
A tbl with an approximate fraction of rows as set by \cod{size} parameter.
}
\description{
Random sample from a xdf object
}
\examples{
mtcarsXdf <- rxDataStep(mtcars )
sample_frac(mtcarsXdf, .1)
}
\seealso{
\code{\link[dplyr]{sample_frac}} in package dplyr
}

