% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/joins_xdf.R
\name{semi_join.RxFileData}
\alias{anti_join.RxFileData}
\alias{intersect.RxFileData}
\alias{semi_join.RxFileData}
\alias{setdiff.RxFileData}
\alias{union.RxFileData}
\title{Unsupported joins/set operations}
\usage{
\method{semi_join}{RxFileData}(x, y, ...)

\method{anti_join}{RxFileData}(x, y, ...)

intersect.RxFileData(x, y, ...)

union.RxFileData(x, y, ...)

setdiff.RxFileData(x, y, ...)
}
\arguments{
\item{x,}{y ScaleR data sources, or tbls wrapping the same.}

\item{...}{Other arguments passed on to methods.}
}
\description{
These operations are currently not supported for ScaleR data sources.
}

