% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdir_utils.R
\name{setXdfTblDir}
\alias{getXdfTblDir}
\alias{setXdfTblDir}
\title{Get and set the xdf tbl directory}
\usage{
setXdfTblDir(dir, fileSystem = rxGetFileSystem())

getXdfTblDir(fileSystem = rxGetFileSystem())
}
\arguments{
\item{dir}{Directory to use for saving xdf tbls}

\item{fileSystem}{The filesystem for which to set the tbl directory; can be either "hdfs" or "local". Currently only the local filesystem is supported.}
}
\description{
Get and set the xdf tbl directory
}
\details{
By default, dplyrXdf will save the xdf files it creates to the R temporary directory. This can be a problem if the temporary directory is in a location with limited disk space. To change the directory to another location, use \code{setXdfTblDir}; to view the current location, use \code{getXdfTblDir}.
}
\section{warning}{
 
You should only set the tbl directory at the beginning of an R session. If you change the directory in the middle of a session, dplyrXdf will no longer be able to find its tbls, which means they may become inaccessible.
}

