% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_mssql.R
\name{db_compute.Microsoft SQL Server}
\alias{db_compute.Microsoft SQL Server}
\alias{db_compute}
\alias{db_save_query}
\alias{db_save_query.Microsoft SQL Server}
\title{SQL Server methods for dbplyr}
\usage{
"db_compute.Microsoft SQL Server"(con, table, sql, temporary = TRUE,
  unique_indexes = list(), indexes = list(), ...)

\method{db_save_query}{Microsoft SQL Server}(con, sql, name, temporary = TRUE,
  ...)
}
\arguments{
\item{con}{A DBI connection}

\item{table, name}{A table name}

\item{sql}{A query}

\item{temporary}{Whether to generate a temporary table (which will be automatically prefixed with \code{'##'})}
}
\description{
Saving tables on SQL Server with dplyr requires creating some custom methods. These are automatically called by \code{compute}, \code{collect} and \code{copy_to}.
}
