% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_xdf.R
\name{summarise_.RxFileData}
\alias{summarise}
\alias{summarise_.RxFileData}
\alias{summarize}
\title{Summarise multiple values to a single value}
\usage{
\method{summarise_}{RxFileData}(.data, ..., .dots)
}
\arguments{
\item{.data}{A tbl for an Xdf data source; or a raw Xdf data source.}

\item{...}{Name-value pairs of summary functions like \code{\link{min}()}, \code{\link{mean}()}, \code{\link{max}()} etc.}

\item{.dots}{Used to work around non-standard evaluation. See the dplyr vignettes for details. See "Details" below for how to change the summarisation method.}
}
\description{
Summarise multiple values to a single value
}
\details{
There are 5 possible methods for doing the summarisation. To choose which method is used, specify a \code{.method} argument in the call to \code{summarise}, with a number from 1 to 5.
\enumerate{
 \item use \code{\link[RevoScaleR]{rxCube}}, cbind data frames together: only \code{n()}, \code{mean()}, \code{sum()} supported, grouped data only (fast)
 \item use \code{\link[RevoScaleR]{rxSummary}}, cbind data frames together: stats in rxSummary supported (fast)
 \item as 2), but build classification levels by pasting the grouping variable(s) together (moderately fast)
 \item split into multiple Xdfs by group, run \code{dplyr::summarise} on each, rbind xdfs together: arbitrary stats supported (slow)
 \item split into multiple Xdfs by group, run \code{rxSummary} on each, rbind xdfs together: stats in rxSummary supported (slowest, most scalable)
}
The default method is 1 if the data is grouped and the requested summary statistics are supported by \code{rxCube}; otherwise 2 if the requested statistics are supported by \code{rxSummary}; otherwise 4. Method 3 is supplied for the case where the product of factor levels for the grouping variables exceeds 2^32 - 1)
}
\seealso{
\code{\link[dplyr]{summarise}} in package dplyr, \code{\link[RevoScaleR]{rxCube}}, \code{\link[RevoScaleR]{rxSummary}}
}

