% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdir_utils.R
\name{set_dplyrxdf_dir}
\alias{get_dplyrxdf_dir}
\alias{set_dplyrxdf_dir}
\title{Get and set the xdf tbl directory}
\usage{
set_dplyrxdf_dir(path, fileSystem = rxGetFileSystem())

get_dplyrxdf_dir(fileSystem = rxGetFileSystem())
}
\arguments{
\item{path}{Location in which to save xdf tbls. If missing, defaults to the R temporary directory.}

\item{fileSystem}{The filesystem for which to set or get the tbl directory; can be either "hdfs" or "native". Currently only the native (local) filesystem is supported.}
}
\description{
By default, dplyrXdf will save the xdf files it creates in the R temporary directory. This can be a problem if it is in a location with limited disk space. Use \code{set_dplyrxdf_dir} to change the xdf tbl directory, and \code{get_dplyrxdf_dir} to view it.
}
\details{
If \code{path} is supplied, \code{set_dplyrxdf_dir} creates a new directory (with a unique name) located \emph{under} \code{path}. This ensures that the files managed by dplyrXdf are properly isolated from the rest of the filesystem.
}
\seealso{
\code{\link{rxGetFileSystem}}, \code{\link{rxSetFileSystem}}
}

