% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdf_utils.R
\name{copy_xdf}
\alias{copy_xdf}
\alias{move_xdf}
\alias{rename_xdf}
\alias{delete_xdf}
\title{Utility functions for working with Xdf files}
\usage{
copy_xdf(src, dest)

move_xdf(src, dest)

rename_xdf(src, dest)

delete_xdf(xdf)
}
\arguments{
\item{src}{For \code{copy_xdf}, \code{move_xdf} and \code{rename_xdf}, an Xdf data source object (\emph{not} a filename).}

\item{dest}{For \code{copy_xdf}, \code{move_xdf} and \code{rename_xdf}, a character string giving the destination file. Note that for \code{rename_xdf}, this should only be a base name, not a full path.}

\item{xdf}{For \code{delete_xdf}, an Xdf data source object.}
}
\value{
\code{copy_xdf}, \code{move_xdf} and \code{rename_xdf} return an Xdf data source object pointing to the new file location. \code{delete_xdf} returns TRUE if the delete operation succeeded, FALSE otherwise.
}
\description{
Copy, move, rename and delete an Xdf file.
}
\details{
The \code{copy_xdf} function copies the Xdf file given by \code{src} to the location specified by \code{dest}, possibly renaming it as well; \code{move_xdf} moves the file. \code{rename_xdf} does a strict rename (the location of the file is unchanged, only its name). \code{delete_xdf} deletes the file given by \code{xdf}.

These are utility functions for working with the files/directories where the data for an Xdf data source is stored. They make use of low-level OS functionality, so should be more efficient that running \code{rxDataStep}. They work with both standard and composite Xdf files, and with data stored both in the native filesystem and in HDFS.
}
\examples{
mtx <- as_xdf(mtcars, overwrite=TRUE)

mtx2 <- copy_xdf(mtx, "mtcars2.xdf")
head(mtx2)
delete_xdf(mtx2)
file.exists("mtcars2.xdf")  # FALSE

mtx2 <- rename_xdf(mtx, "mtcars2.xdf")
file.exists("mtcars.xdf")   # FALSE
file.exists("mtcars2.xdf")  # TRUE
}
\seealso{
\code{\link{file.copy}}, \code{\link{file.rename}}, \code{\link{unlink}},
\code{\link{rxHadoopCopy}}, \code{\link{rxHadoopMove}}, \code{\link{rxHadoopRemove}}, \code{\link{rxHadoopRemoveDir}}
}
