% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange.RxXdfData}
\alias{arrange.RxXdfData}
\alias{arrange}
\alias{arrange.RxFileData}
\title{Arrange the rows in an Xdf file}
\usage{
\method{arrange}{RxXdfData}(.data, ..., .by_group = FALSE,
  .outFile = tbl_xdf(.data), .rxArgs)

\method{arrange}{RxFileData}(.data, ...)
}
\arguments{
\item{...}{List of unquoted variable names. Use \code{desc} to sort in descending order.}

\item{.by_group}{If TRUE, will sort first by grouping variables.}

\item{.outFile}{Output format for the returned data. If not supplied, create an xdf tbl; if \code{NULL}, return a data frame; if a character string naming a file, save an Xdf file at that location.}

\item{.rxArgs}{A list of RevoScaleR arguments. See \code{\link{rxArgs}} for details.}

\item{data}{An Xdf data source, tbl, or other RevoScaleR file data source.}
}
\value{
An object representing the sorted data. This depends on the \code{.outFile} argument: by default, it will be an xdf tbl object; if \code{NULL}, a data frame; and if a filename, an Xdf data source referencing a file saved to that location.
}
\description{
Arrange the rows in an Xdf file
}
\details{
The underlying RevoScaleR function is \code{rxSort}. This has many sorting options, including removing duplicated keys, adding a column of frequency counts, and so on.
}
\seealso{
\code{\link{rxSort}}, \code{\link[dplyr]{arrange}} in package dplyr
}
