% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_df_methods.R
\name{as.data.frame.RxXdfData}
\alias{$.RxFileData}
\alias{[[.RxFileData}
\alias{as.data.frame.RxFileData}
\alias{as.data.frame.RxXdfData}
\title{Convert a data source or tbl to a data frame}
\usage{
\method{as.data.frame}{RxXdfData}(x, maxRowsByCols = NULL, ...)

\method{as.data.frame}{RxFileData}(x, maxRowsByCols = NULL, ...)

\method{$}{RxFileData}(x, name)

\method{[[}{RxFileData}(x, name, maxRowsByCols = NULL, ...)
}
\arguments{
\item{x}{A data source object, or tbl wrapping the same.}

\item{maxRowsByCols}{the maximum dataset size to convert, expressed in terms of rows times columns. Defaults to NULL, meaning no maximum.}

\item{...}{Other arguments to \code{rxDataStep}.}

\item{name}{The name of a column to extract from a data source object}
}
\description{
Convert a data source or tbl to a data frame
}
\details{
These are simple wrappers around \code{\link[RevoScaleR]{rxDataStep}}, with the check on the maximum table size turned off. You should ensure that you have enough memory for your data.

\code{as.data.frame} converts a data source object (typically an xdf file, but can also be any data source type that \code{rxDataStep} supports) into a data frame. The \code{$} and \code{[[} methods extract a single column from a data source, as a vector.
}

