% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_xdf.R
\name{as.data.frame.RxXdfData}
\alias{as.data.frame.RxFileData}
\alias{as.data.frame.RxXdfData}
\title{Convert a data source or tbl to a data frame}
\usage{
\method{as.data.frame}{RxXdfData}(x, maxRowsByCols = NULL, ...)

\method{as.data.frame}{RxFileData}(x, maxRowsByCols = NULL, ...)
}
\arguments{
\item{x}{A data source object, or tbl wrapping the same.}

\item{maxRowsByCols}{the maximum dataset size to convert, expressed in terms of rows times columns. Defaults to NULL, meaning no maximum.}

\item{...}{Other arguments to \code{rxDataStep}.}
}
\description{
Convert a data source or tbl to a data frame
}
\details{
This is a simple wrapper around \code{\link[RevoScaleR]{rxDataStep}}, with the check on the maximum table size turned off. You should ensure that you have enough memory for your data.
}

