% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_df_methods.R
\name{as.data.frame.RxXdfData}
\alias{as.data.frame.RxXdfData}
\alias{as.data.frame}
\alias{as.data.frame.RxDataSource}
\alias{$.RxFileData}
\alias{[[.RxFileData}
\alias{pull.RxFileData}
\title{Convert a data source or tbl to a data frame}
\usage{
\method{as.data.frame}{RxXdfData}(x, maxRowsByCols = NULL, row.names = NULL,
  optional = TRUE, ...)

\method{as.data.frame}{RxDataSource}(x, maxRowsByCols = NULL,
  row.names = NULL, optional = TRUE, ...)

\method{$}{RxFileData}(x, name)

\method{[[}{RxFileData}(x, name, maxRowsByCols = NULL, ...)

\method{pull}{RxFileData}(.data, var = -1)
}
\arguments{
\item{x}{A data source object, or tbl wrapping the same.}

\item{maxRowsByCols}{the maximum dataset size to convert, expressed in terms of rows times columns. Defaults to NULL, meaning no maximum.}

\item{row.names, optional}{For compatibility with the \code{as.data.frame} generic. Not used.}

\item{...}{Other arguments to \code{rxDataStep}.}

\item{name}{The name of a column to extract from a data source object}
}
\description{
Convert a data source or tbl to a data frame
}
\details{
These are simple wrappers around \code{\link[RevoScaleR]{rxDataStep}}, with the check on the maximum table size turned off. You should ensure that you have enough memory for your data.

\code{as.data.frame} converts a data source object (typically an Xdf file, but can also be any data source type that \code{rxDataStep} supports) into a data frame. The \code{pull}, \code{$} and \code{[[} methods extract a single column from a data source, as a vector.
}
\examples{
mtx <- as_xdf(mtcars, overwrite=TRUE)
mtx$mpg
mtx[["mpg"]]
pull(mtx, mpg)
as.data.frame(mtx)
as_data_frame(mtx)  # returns a tbl_df
}
\seealso{
\code{\link[base]{as.data.frame}}, \code{\link{collect}}
}
