% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_to_hdfs.R
\name{copy_to.RxHdfsFileSystem}
\alias{copy_to.RxHdfsFileSystem}
\alias{copy_to}
\alias{copy_to_hdfs}
\title{Upload a dataset to HDFS}
\usage{
\method{copy_to}{RxHdfsFileSystem}(dest, df, name = NULL, ...)

copy_to_hdfs(...)
}
\arguments{
\item{dest}{The destination source: an object of class \code{\link{RxHdfsFileSystem}}.}

\item{df}{A dataset: can be a filename, an Xdf data source object, another RevoScaleR data source, or anything that can be coerced to a data frame.}

\item{name}{The filename, optionally including the path, for the uploaded Xdf file. Defaults to the name of the source dataset, in the user's HDFS home directory.}

\item{...}{Further arguments to \code{\link{rxHadoopCommand}}.}
}
\value{
An Xdf data source object pointing to the uploaded data.
}
\description{
Upload a dataset to HDFS
}
\details{
This is the RevoScaleR HDFS method for the dplyr \code{\link[dplyr]{copy_to}} function, for uploading data to a remote database/src. The method should work with any RevoScaleR data source, or with any R object that can be converted into a data frame. If the data is not already in Xdf format, it is first imported into Xdf, and then uploaded.

The code will handle both the cases where you are logged into the edge node of a Hadoop/Spark cluster, and if you are a remote client. For the latter case, the uploading is a two-stage process: the data is first transferred to the native filesystem of the edge node, and then copied from the edge node into HDFS.

The \code{copy_to_hdfs} function is a simple wrapper that avoids having to create an explicit filesystem object.
}
\section{Note on composite Xdf}{

There are actually two kinds of Xdf files: standard and _composite_. A composite Xdf file is a directory containing multiple data and metadata files, which the RevoScaleR functions treat as a single dataset. Xdf files in HDFS must be composite in order to work properly; \code{copy_to} will convert an existing Xdf file into composite, if it's not already in that format. Non-Xdf datasets (data frames and other RevoScaleR data sources, such as text files) will similarly be uploaded as composite.
}

\seealso{
\code{\link{rxHadoopCopyFromClient}}, \code{\link{rxHadoopCopyFromLocal}},
\code{\link{collect}} and \code{\link{compute}} for downloading data from HDFS,
\code{\link{as_xdf}}, \code{\link{as_composite_xdf}}
}
