% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_xdf.R
\name{distinct_.RxFileData}
\alias{distinct}
\alias{distinct_}
\alias{distinct_.RxFileData}
\alias{distinct_.grouped_tbl_xdf}
\title{Select distinct/unique rows}
\usage{
\method{distinct_}{RxFileData}(.data, ..., .output, .rxArgs, .dots,
  .keep_all = FALSE)

\method{distinct_}{grouped_tbl_xdf}(.data, ..., .output, .rxArgs, .dots,
  .keep_all = FALSE)
}
\arguments{
\item{.data}{A tbl for an Xdf data source; or a raw Xdf data source.}

\item{...}{Variables to use for determining uniqueness. If left blank, all variables in \code{.data} are used to determine uniqueness.}

\item{.output}{Output format for the returned data. If not supplied, create an xdf tbl; if \code{NULL}, return a data frame; if a character string naming a file, save an Xdf file at that location.}

\item{.rxArgs}{A list of RevoScaleR arguments. See \code{\link{rxArgs}} for details.}

\item{.dots}{Used to work around non-standard evaluation. See the dplyr vignettes for details.}

\item{.keep_all}{If \code{TRUE}, keep all variables in the dataset; otherwise, only keep variables used in defining uniqueness. Only used if dplyr version 0.5 or greater is installed.}
}
\value{
An object representing the unique rows. This depends on the \code{.output} argument: if missing, it will be an xdf tbl object; if \code{NULL}, a data frame; and if a filename, an Xdf data source referencing a file saved to that location.
}
\description{
Select distinct/unique rows
}
\details{
This verb calls \code{dplyr::distinct} on each chunk in an Xdf file. The individual data frames are \code{rbind}ed together and \code{dplyr::distinct} is called on the overall result. This may be slow if there are many chunks in the file; and the operation will be limited by memory if the number of distinct rows is large.
}
\seealso{
\code{\link[dplyr]{distinct}} in package dplyr
}

