% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distinct_xdf.R
\name{distinct_.RxFileData}
\alias{distinct}
\alias{distinct_.RxFileData}
\alias{distinct_.grouped_tbl_xdf}
\title{Select distinct/unique rows}
\usage{
\method{distinct_}{RxFileData}(.data, ..., .dots)

\method{distinct_}{grouped_tbl_xdf}(.data, ..., .dots)
}
\arguments{
\item{.data}{A tbl for an Xdf data source; or a raw Xdf data source.}

\item{...}{Variables to use for determining uniqueness.}

\item{.dots}{Used to work around non-standard evaluation. See the dplyr vignettes for details.}
}
\description{
Select distinct/unique rows
}
\details{
To sort a grouped Xdf tbl, \code{distinct} creates a factor variable \code{.group.} containing all observed combinations of the grouping variables. It then uses \code{\link{rxSplit}} to split the data into multiple files by this factor, and calls \code{\link[dplyr]{distinct}} on each. The individual data frames are \code{rbind}ed together and \code{distinct} is called on the overall result. This ensures that the groups will be appropriately generated regardless of the types of the grouping variables. Note however that this may be slow if you have a large number of groups; and the operation will be limited by memory if the number of distinct rows is large.
}
\seealso{
\code{\link[dplyr]{distinct}} in package dplyr
}

