% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{dplyrxdf_options}
\alias{dplyrxdf_options}
\title{Get or set dplyrXdf options}
\usage{
dplyrxdf_options(...)
}
\arguments{
\item{...}{Options to define, using \code{name = value}. Valid options are listed under "Details" below.}
}
\value{
Returns a list of the option values before any changes are carried out.
}
\description{
Get or set dplyrXdf options
}
\details{
Use this function to get or set various options that control dplyrXdf's functionality.
\itemize{
 \item \code{useExecBy}: whether to do by-group processing with \code{\link[RevoScaleR]{rxExecBy}}. Defaults to TRUE if Microsoft R Server 9.1 or later is installed. Requires MRS 9.1.
 \item \code{localWorkDir}: working directory to store xdf tbl files. It is recommended to use \code{\link{get_dplyrxdf_dir}} and \code{\link{set_dplyrxdf_dir}} for getting and setting this option.
 \item \code{hdfsWorkDir}: working directory to store xdf tbl files in HDFS. Not currently used; for future expansion.
 \item \code{rowsPerRead}: default number of rows per chunk when writing xdf files.
 \item \code{dplyrVersion}: version of dplyr installed.
}
}
\seealso{
\code{\link{options}}
}
