% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_xdf.R
\name{filter.RxFileData}
\alias{filter.RxFileData}
\alias{filter}
\title{Subset a data source by rows}
\usage{
\method{filter}{RxFileData}(.data, ..., .outFile = tbl_xdf(.data), .rxArgs)
}
\arguments{
\item{.data}{A data source.}

\item{...}{Expressions to filter by.}

\item{.outFile}{Output format for the returned data. If not supplied, create an xdf tbl; if \code{NULL}, return a data frame; if a character string naming a file, save an Xdf file at that location.}

\item{.rxArgs}{A list of RevoScaleR arguments. See \code{\link{rxArgs}} for details.}

\item{.dots}{Used to work around non-standard evaluation. See the dplyr vignettes for details.}
}
\value{
An object representing the filtered data. This depends on the \code{.outFile} argument: if missing, it will be an xdf tbl object; if \code{NULL}, a data frame; and if a filename, an Xdf data source referencing a file saved to that location.
}
\description{
Subset a data source by rows
}
\seealso{
\code{\link[dplyr]{filter}} in package dplyr
}
