% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/group_by_xdf.R
\name{group_by_.RxFileData}
\alias{group_by}
\alias{group_by_}
\alias{group_by_.RxFileData}
\title{Group an Xdf file by one or more variables}
\usage{
\method{group_by_}{RxFileData}(.data, ..., .dots, add = FALSE)
}
\arguments{
\item{.data}{An Xdf file or a tbl wrapping the same.}

\item{...}{Variables to group by.}

\item{.dots}{Used to work around non-standard evaluation.}

\item{add}{If FALSE (the default), \code{group_by} will ignore existing groups. If TRUE, add grouping variables to existing groups.}
}
\description{
Group an Xdf file by one or more variables
}
\details{
\code{group_by} itself does not do any data processing; it only sets up the necessary metadata for verbs accepting grouped tbls to generate the groups.

Most verbs that accept grouped data as input will use \code{\link{rxSplit}} to split the data into multiple Xdf files, and then process each file separately. The exception is \code{\link[dplyrXdf]{summarise}}, which allows a range of options in how to treat groups.
}
\seealso{
\code{\link[dplyr]{group_by}}
}

