% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins_xdf.R
\name{join}
\alias{anti_join}
\alias{anti_join.RxFileData}
\alias{full_join}
\alias{full_join.RxFileData}
\alias{inner_join}
\alias{inner_join.RxFileData}
\alias{join}
\alias{left_join}
\alias{left_join.RxFileData}
\alias{right_join}
\alias{right_join.RxFileData}
\alias{semi_join}
\alias{semi_join.RxFileData}
\title{Join two data sources together}
\usage{
\method{left_join}{RxFileData}(x, y, by = NULL, copy = FALSE, .output,
  .rxArgs, ...)

\method{right_join}{RxFileData}(x, y, by = NULL, copy = FALSE, .output,
  .rxArgs, ...)

\method{inner_join}{RxFileData}(x, y, by = NULL, copy = FALSE, .output,
  .rxArgs, ...)

\method{full_join}{RxFileData}(x, y, by = NULL, copy = FALSE, .output,
  .rxArgs, ...)

\method{semi_join}{RxFileData}(x, y, by = NULL, copy = FALSE, .output,
  .rxArgs, ...)

\method{anti_join}{RxFileData}(x, y, by = NULL, copy = FALSE, .output,
  .rxArgs, ...)
}
\arguments{
\item{x, }{y Data sources to join.}

\item{by}{Character vector of variables to join by. See \code{\link[dplyr]{join}} for details.}

\item{copy}{If the data sources are not stored in the same system, whether to copy y to x's location. Not currently used.}

\item{.output}{Output format for the returned data. If not supplied, create an xdf tbl; if \code{NULL}, return a data frame; if a character string naming a file, save an Xdf file at that location.}

\item{.rxArgs}{A list of RevoScaleR arguments. See \code{\link{rxArgs}} for details.}

\item{...}{Not currently used.}
}
\value{
An object representing the joined data. This depends on the \code{.output} argument: if missing, it will be an xdf tbl object; if \code{NULL}, a data frame; and if a filename, an Xdf data source referencing a file saved to that location.
}
\description{
Join two data sources together
}
\details{
Currently joining only supports the local file system.
}
\seealso{
\code{\link[dplyr]{join}} in package dplyr
}

