% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins_unsupported.R, R/union_xdf.R
\docType{import}
\name{reexports}
\alias{intersect}
\alias{reexports}
\alias{setdiff}
\alias{setequal}
\alias{union}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{intersect}}, \code{\link[dplyr]{setdiff}}, \code{\link[dplyr]{setequal}}, \code{\link[dplyr]{union}}}
}}
\keyword{internal}

