% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_xdf.R
\name{select_.RxFileData}
\alias{select}
\alias{select_}
\alias{select_.RxFileData}
\title{Select columns in a data source}
\usage{
\method{select_}{RxFileData}(.data, ..., .output, .rxArgs, .dots)
}
\arguments{
\item{.data}{A data source.}

\item{...}{Unquoted variables to select.}

\item{.output}{Output format for the returned data. If not supplied, create an xdf tbl; if \code{NULL}, return a data frame; if a character string naming a file, save an Xdf file at that location.}

\item{.rxArgs}{A list of RevoScaleR arguments. See \code{\link{rxArgs}} for details.}

\item{.dots}{Used to work around non-standard evaluation. See the dplyr vignettes for details.}
}
\value{
An object representing the transformed data. This depends on the \code{.output} argument: if missing, it will be an xdf tbl object; if \code{NULL}, a data frame; and if a filename, an Xdf data source referencing a file saved to that location.
}
\description{
Select columns in a data source
}
\details{
All the special functions mentioned in the documentation for \code{\link[dplyr]{select}} will also work with dplyrXdf. Note that renaming a variable is \emph{not} supported in dplyrXdf \code{select}. If you want to do this, follow the \code{select} with a \code{rename}; the latter is very fast, as it only modifies the metadata portion of an xdf file, not the data itself.
}
\seealso{
\code{\link[dplyr]{select}} in package dplyr
}

