% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins_unsupported.R, R/union_xdf.R
\name{intersect.RxFileData}
\alias{intersect}
\alias{intersect.RxFileData}
\alias{setdiff}
\alias{setdiff.RxFileData}
\alias{setequal}
\alias{setequal.RxFileData}
\alias{setops}
\alias{union}
\alias{union.RxFileData}
\alias{union_all}
\alias{union_all.RxFileData}
\title{Set operations on data sources}
\usage{
\method{intersect}{RxFileData}(x, y, ...)

\method{setdiff}{RxFileData}(x, y, ...)

\method{setequal}{RxFileData}(x, y, ...)

\method{union_all}{RxFileData}(x, y, .outFile, .rxArgs, ...)

\method{union}{RxFileData}(x, y, .outFile, .rxArgs, ...)
}
\arguments{
\item{x, }{y Data sources.}

\item{...}{Not currently used.}

\item{.outFile}{Output format for the returned data. If not supplied, create an xdf tbl; if \code{NULL}, return a data frame; if a character string naming a file, save an Xdf file at that location.}

\item{.rxArgs}{A list of RevoScaleR arguments. See \code{\link{rxArgs}} for details.}
}
\value{
An object representing the joined data. This depends on the \code{.outFile} argument: if missing, it will be an xdf tbl object; if \code{NULL}, a data frame; and if a filename, an Xdf data source referencing a file saved to that location.
}
\description{
Set operations on data sources
}
\details{
Currently, only \code{union} and \code{union_all} are supported for RevoScaleR data sources. The code uses \code{rxDataStep(append="rows")} to do the union; this can be much faster than using \code{rxMerge(type="union")}.

For technical reasons, \code{union_all} is only exported as a method for RevoScaleR objects if you have dplyr version 0.5 or higher. If an earlier version of dplyr is installed, you can still call it but you'll have to specify the full function name: \code{union_all.RxFileData}. If this doesn't make sense to you, just assume that \code{union_all} requires dplyr 0.5.
}
\seealso{
\code{\link[dplyr]{setops}} in package dplyr
}

