% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins_xdf.R
\name{intersect.RxFileData}
\alias{intersect.RxFileData}
\alias{setdiff.RxFileData}
\alias{union.RxFileData}
\title{Unsupported joins/set operations}
\usage{
intersect.RxFileData(x, y, ...)

union.RxFileData(x, y, ...)

setdiff.RxFileData(x, y, ...)
}
\arguments{
\item{x, }{y ScaleR data sources, or tbls wrapping the same.}

\item{...}{Other arguments passed on to methods.}
}
\description{
These operations are currently not supported for ScaleR data sources.
}

