% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins_unsupported.R, R/union_xdf.R
\name{intersect.RxFileData}
\alias{intersect.RxFileData}
\alias{setdiff.RxFileData}
\alias{setequal.RxFileData}
\alias{setops}
\alias{union}
\alias{union_all}
\alias{intersect}
\alias{setdiff}
\alias{setequal}
\alias{union_all.RxFileData}
\alias{union.RxFileData}
\title{Set operations on data sources}
\usage{
\method{intersect}{RxFileData}(x, y, ...)

\method{setdiff}{RxFileData}(x, y, ...)

\method{setequal}{RxFileData}(x, y, ...)

\method{union_all}{RxFileData}(x, y, .outFile = tbl_xdf(x), .rxArgs)

\method{union}{RxFileData}(x, y, .outFile = tbl_xdf(x), .rxArgs, ...)
}
\arguments{
\item{x, y}{Data sources.}

\item{...}{Not currently used.}

\item{.outFile}{Output format for the returned data. If not supplied, create an xdf tbl; if \code{NULL}, return a data frame; if a character string naming a file, save an Xdf file at that location.}

\item{.rxArgs}{A list of RevoScaleR arguments. See \code{\link{rxArgs}} for details.}
}
\value{
An object representing the joined data. This depends on the \code{.outFile} argument: if missing, it will be an xdf tbl object; if \code{NULL}, a data frame; and if a filename, an Xdf data source referencing a file saved to that location.
}
\description{
Set operations on data sources
}
\details{
Currently, only \code{union} and \code{union_all} are supported for RevoScaleR data sources. The code uses \code{rxDataStep(append="rows")} to do the union; this can be much faster than using \code{rxMerge(type="union")}.
}
\seealso{
\code{\link[dplyr]{setops}} in package dplyr
}
