% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_utils.R
\docType{methods}
\name{tbl_vars.RxFileData}
\alias{tbl_vars.RxFileData}
\alias{tbl_vars}
\alias{tbl_types}
\alias{names,tbl_xdf-method}
\alias{names<-,tbl_xdf,character-method}
\alias{names<-,grouped_tbl_xdf,character-method}
\title{Get the variable names or types for a RevScaleR data source}
\usage{
\method{tbl_vars}{RxFileData}(x)

tbl_types(x)

\S4method{names}{tbl_xdf}(x)

\S4method{names}{tbl_xdf,character}(x) <- value

\S4method{names}{grouped_tbl_xdf,character}(x) <- value
}
\arguments{
\item{x}{A data source object, or tbl wrapping the same.}
}
\value{
For \code{tbl_vars}, a character vector of variable names; for \code{tbl_types}, a named vector giving the types of each variable.
}
\description{
Get the variable names or types for a RevScaleR data source
}
\details{
These are simple wrappers around the \code{names} and \code{rxGetVarInfo} functions.
}
\seealso{
\code{\link{RxXdfData}}, \code{\link{RxTextData}}, \code{\link{RxSasData}}, \code{\link{RxSpssData}},
\code{\link{rxGetInfo}}, \code{\link{rxGetVarInfo}}
}
